+++ R/autoplot-density.R
@@ -88,7 +88,8 @@ tidy_autoplot <- function(.data, .plot_type = "density", .line_size = .5,
     "tidy_zero_truncated_binomial", "tidy_zero_truncated_negative_binomial",
     "tidy_pareto_single_parameter", "tidy_pareto", "tidy_inverse_pareto",
     "tidy_generalized_pareto","tidy_paralogistic", "tidy_inverse_exponential",
-    "tidy_inverse_gamma","tidy_inverse_weibull","tidy_burr","tidy_inverse_burr"
+    "tidy_inverse_gamma","tidy_inverse_weibull","tidy_burr","tidy_inverse_burr",
+    "tidy_inverse_gaussian"
   )) {
     rlang::abort("The data passed must come from a `tidy_` distribution function.")
   }
@@ -173,6 +174,10 @@ tidy_autoplot <- function(.data, .plot_type = "density", .line_size = .5,
              "Scale: ", atb$.scale)
     } else if (atb$tibble_type == "tidy_inverse_exponential"){
       paste0("Rate: ", atb$.rate, " - Scale: ", atb$.scale)
+    } else if (atb$tibble_type == "tidy_inverse_gaussian"){
+      paste0("Mean: ", atb$.mean, " - ",
+             "Shape: ", atb$.shape, " - ",
+             "Dispersion: ", atb$.dispersion)
     }
   )